feat(note&user):realize note public api and user findById api

+++ oursphere-note/oursphere-note-biz/src/main/java/com/theoyu/oursphere/note/biz/model/entity/NotePO.java
@@ -25,19 +25,19 @@ public class NotePO {
 
     private Boolean isTop;
 
-    private Byte type;
+    private Integer  type;
 
     private String imgUris;
 
     private String videoUri;
 
-    private Byte visible;
+    private Integer  visible;
 
     private LocalDateTime  createTime;
 
     private LocalDateTime  updateTime;
 
-    private Byte status;
+    private Integer  status;
 
     private String contentUuid;
